+++ bmc-core/src/main/java/com/oracle/bmc/core/model/CreateSubnetDetails.java
@@ -135,23 +135,22 @@ public class CreateSubnetDetails {
     /**
      * A DNS label for the subnet, used in conjunction with the VNIC's hostname and
      * VCN's DNS label to form a fully qualified domain name (FQDN) for each VNIC
-     * within this subnet (e.g., `bminstance-1.subnet-123.vcn-1.oraclevcn.com`).
-     * Must be unique within the VCN and comply with
-     * [RFC 952](https://tools.ietf.org/html/rfc952) and
-     * [RFC 1123](https://tools.ietf.org/html/rfc1123). The value cannot be changed.
+     * within this subnet (e.g., `bminstance-1.subnet123.vcn1.oraclevcn.com`).
+     * Must be an alphanumeric string that begins with a letter and is unique within the VCN.
+     * The value cannot be changed.
      * <p>
-     * This value must be set if you want to use the VCN Resolver to resolve the
+     * This value must be set if you want to use the Internet and VCN Resolver to resolve the
      * hostnames of instances in the subnet. It can only be set if the VCN itself
      * was created with a DNS label.
      * <p>
      * For more information, see
      * [DNS in Your Virtual Cloud Network](https://docs.us-phoenix-1.oraclecloud.com/Content/Network/Concepts/dns.htm).
      * <p>
-     * Example: `subnet-123`
+     * Example: `subnet123`
      *
      **/
     @JsonProperty("dnsLabel")
-    @Size(min = 1, max = 63)
+    @Size(min = 1, max = 15)
     String dnsLabel;
 
     /**