+++ bmc-core/src/main/java/com/oracle/bmc/core/model/Instance.java
@@ -171,6 +171,11 @@ public class Instance {
      * following iSCSI IP address: 169.254.0.2, and boot volume IQN:
      * iqn.2015-02.oracle.boot.
      * <p>
+     * For more information about the Bring Your Own Image feature of
+     * Oracle Bare Metal Cloud Services, see [Bring Your Own Custom
+     * Image: RHEL 7.x & Derivatives (CentOS, Oracle Linux)]
+     * (/Content/General/Reference/aqswhitepapers.htm).
+     *
      * For more information about iPXE, see http://ipxe.org.
      *
      **/